factored out common visualization methods
