+++ perception_tools/iterative_closest_point.py
@@ -7,86 +7,6 @@ import meshcat.transformations as tf
 
 from sklearn.neighbors import NearestNeighbors
 
-# TODO(kmuhlrad): move to separate library
-def MakeMeshcatColorArray(N, r, g, b):
-    """Constructs a color array to visualize a point cloud in meshcat.
-
-    @param N int. Number of points to generate. Must be >= number of points in
-        the point cloud to color.
-    @param r float. The red value of the points, 0.0 <= r <= 1.0.
-    @param g float. The green value of the points, 0.0 <= g <= 1.0.
-    @param b float. The blue value of the points, 0.0 <= b <= 1.0.
-
-    @return 3xN numpy array of the same color.
-    """
-    color = np.zeros((3, N))
-    color[0, :] = r
-    color[1, :] = g
-    color[2, :] = b
-
-    return color
-
-
-def VisualizeICP(meshcat_vis, scene, model, X_MS):
-    """
-    Visualizes the ground truth (red), observation (blue), and transformed
-    (yellow) point clouds in meshcat.
-
-    @param meschat_vis An instance of a meshcat visualizer.
-    @param scene An Nx3 numpy array representing the scene point cloud.
-    @param model An Mx3 numpy array representing the model point cloud.
-    @param X_GO A 4x4 numpy array of the homogeneous transformation from the
-            scene point cloud to the model point cloud.
-    """
-
-    meshcat_vis['model'].delete()
-    meshcat_vis['observations'].delete()
-    meshcat_vis['transformed_observations'].delete()
-
-    # Make meshcat color arrays.
-    N = scene.shape[0]
-    M = model.shape[0]
-
-    red = MakeMeshcatColorArray(M, 0.5, 0, 0)
-    blue = MakeMeshcatColorArray(N, 0, 0, 0.5)
-    yellow = MakeMeshcatColorArray(N, 1, 1, 0)
-
-    # Create red and blue meshcat point clouds for visualization.
-    model_meshcat = g.PointCloud(model.T, red, size=0.01)
-    observations_meshcat = g.PointCloud(scene.T, blue, size=0.01)
-
-    meshcat_vis['model'].set_object(model_meshcat)
-    meshcat_vis['observations'].set_object(observations_meshcat)
-
-    # Create a copy of the scene point cloud that is homogenous
-    # so we can apply a 4x4 homogenous transform to it.
-    homogenous_scene = np.ones((N, 4))
-    homogenous_scene[:, :3] = np.copy(scene)
-
-    # Apply the returned transformation to the scene samples to align the
-    # scene point cloud with the ground truth point cloud.
-    transformed_scene = X_MS.dot(homogenous_scene.T)
-
-    # Create a yellow meshcat point cloud for visualization.
-    transformed_scene_meshcat = \
-        g.PointCloud(transformed_scene[:3, :], yellow, size=0.01)
-
-    meshcat_vis['transformed_observations'].set_object(
-        transformed_scene_meshcat)
-
-
-def ClearVis(meshcat_vis):
-    """
-    Removes model, observations, and transformed_observations objects
-    from meshcat.
-
-    @param meschat_vis An instance of a meshcat visualizer.
-    """
-
-    meshcat_vis['model'].delete()
-    meshcat_vis['observations'].delete()
-    meshcat_vis['transformed_observations'].delete()
-
 
 def FindNearestNeighbors(point_cloud_A, point_cloud_B):
     """
