+++ perception_tools/optimization_based_point_cloud_registration.py
@@ -5,80 +5,6 @@ from scipy.optimize import fmin_bfgs
 from sklearn.neighbors import NearestNeighbors
 
 
-# TODO(kmuhlrad): move to separate library
-def MakeMeshcatColorArray(N, r, g, b):
-    """Constructs a color array to visualize a point cloud in meshcat.
-
-    @param N int. Number of points to generate. Must be >= number of points in
-        the point cloud to color.
-    @param r float. The red value of the points, 0.0 <= r <= 1.0.
-    @param g float. The green value of the points, 0.0 <= g <= 1.0.
-    @param b float. The blue value of the points, 0.0 <= b <= 1.0.
-
-    @return 3xN numpy array of the same color.
-    """
-    color = np.zeros((3, N))
-    color[0, :] = r
-    color[1, :] = g
-    color[2, :] = b
-
-    return color.T
-  
-
-def PlotMeshcatPointCloud(meshcat_vis, point_cloud_name, points, colors):
-    """A wrapper function to plot meshcat point clouds.
-
-    Args:
-    @param meshcat_vis An instance of a meshcat visualizer.
-    @param point_cloud_name string. The name of the meshcat point clouds.
-    @param points An Nx3 numpy array of (x, y, z) points.
-    @param colors An Nx3 numpy array of (r, g, b) colors corresponding to
-        points.
-    """
-
-    meshcat_vis[point_cloud_name].set_object(g.PointCloud(points.T, colors.T))
-
-def VisualizeTransform(meshcat_vis, points, transform):
-    """Visualizes the points transformed by transform in yellow.
-
-    Args:
-    @param meschat_vis An instance of a meshcat visualizer.
-    @param points An Nx3 numpy array representing a point cloud.
-    @param transform a 4x4 numpy array representing a homogeneous
-        transformation.
-    """
-
-    # Make meshcat color arrays.
-    N = points.shape[0]
-
-    yellow = MakeMeshcatColorArray(N, 1, 1, 0)
-
-    homogenous_points = np.ones((N, 4))
-    homogenous_points[:, :3] = np.copy(points)
-
-    transformed_points = transform.dot(homogenous_points.T)
-
-    transformed_points_meshcat = \
-        g.PointCloud(transformed_points[:3, :], yellow.T)
-
-    meshcat_vis['transformed_observations'].set_object(
-        transformed_points_meshcat)
-
-
-def ThresholdArray(arr, min_val, max_val):
-    """
-    Finds where the values of arr are between min_val and max_val (inclusive).
-
-    @param arr An (N, ) numpy array containing number values.
-    @param min_val number. The minimum value threshold.
-    @param max_val number. The maximum value threshold.
-
-    @return An (M, ) numpy array of the integer indices in arr with values that
-        are between min_val and max_val.
-    """
-    return np.where(
-        abs(arr - (max_val + min_val) / 2.) < (max_val - min_val) / 2.)[0]
-
 def PoseToTransform(pose):
     """
     Puts an (x, y, sin(theta), cos(theta)) pose into a 4x4 homogenous
